conan_requirements.txt
# Generated for conan_package_tools==0.34.2 bincrafters_package_tools==0.27.5 conan==1.29.0
appdirs==1.4.4
bincrafters-package-tools==0.27.5
bottle==0.12.18
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
conan==1.29.0
conan-package-tools==0.34.2
deprecation==2.0.7
distlib==0.3.1
distro==1.5.0
fasteners==0.15
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
node-semver==0.6.1
packaging==20.4
patch-ng==1.17.4
pluginbase==1.0.0
Pygments==2.7.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
six==1.14.0
tabulate==0.8.7
tqdm==4.49.0
urllib3==1.25.10
virtualenv==20.0.31
zipp==3.1.0
